Merge base develop to SparkDev develop

+++ Rock/Services/NuGet/RockProjectManager.cs
@@ -1,11 +1,11 @@
 ﻿// <copyright>
-// Copyright 2013 by the Spark Development Network
+// Copyright by the Spark Development Network
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
+// Licensed under the Rock Community License (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-// http://www.apache.org/licenses/LICENSE-2.0
+// http://www.rockrms.com/license
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -96,12 +96,23 @@ namespace Rock.Services.NuGet
 
             try
             {
-                Directory.Delete( packageRestorePath, recursive: true );
+                try
+                {
+                    Directory.Delete( packageRestorePath, recursive: true );
+                }
+                catch ( IOException )
+                {
+                    // try one more time
+                    System.Threading.Thread.Sleep( 10 );
+                    if ( Directory.Exists( packageRestorePath ) )
+                    {
+                        Directory.Delete( packageRestorePath, recursive: true );
+                    }
+                }
             }
             catch ( Exception ex )
             {
-                ExceptionLogService.LogException( ex, System.Web.HttpContext.Current );
-                ExceptionLogService.LogException( new Exception( string.Format( "Unable to delete package restore folder ({0}) after a successful update.", packageRestorePath ) ), System.Web.HttpContext.Current );
+                ExceptionLogService.LogException( new Exception( string.Format( "Note: Unable to delete the temporary package restore folder ({0}) after a successful update.", packageRestorePath ), ex ), System.Web.HttpContext.Current );
             }
         }
